refactor: consumer parameters default input
